Move fanout to the Pair folder
