+++ docs/src/pages/docs/crocks/Either.md
@@ -1235,7 +1235,7 @@ import Either from 'crocks/Either'
 
 import Maybe from 'crocks/Maybe'
 import compose from 'crocks/helpers/compose'
-import fanout from 'crocks/helpers/fanout'
+import fanout from 'crocks/Pair/fanout'
 import identity from 'crocks/combinators/identity'
 import isObject from 'crocks/predicates/isObject'
 import maybeToEither from 'crocks/Either/maybeToEither'
@@ -1302,7 +1302,7 @@ import Result from 'crocks/Result'
 import assign from 'crocks/helpers/assign'
 import compose from 'crocks/helpers/compose'
 import composeK from 'crocks/helpers/composeK'
-import fanout from 'crocks/helpers/fanout'
+import fanout from 'crocks/Pair/fanout'
 import isNumber from 'crocks/predicates/isNumber'
 import liftA2 from 'crocks/helpers/liftA2'
 import map from 'crocks/pointfree/map'
