+++ docs/src/pages/docs/crocks/Maybe.md
@@ -584,7 +584,7 @@ import Maybe from 'crocks/Maybe'
 import compose from 'crocks/helpers/compose'
 import chain from 'crocks/pointfree/chain'
 import curry from 'crocks/helpers/curry'
-import fanout from 'crocks/helpers/fanout'
+import fanout from 'crocks/Pair/fanout'
 import isString from 'crocks/predicates/isString'
 import liftA2 from 'crocks/helpers/liftA2'
 import merge from 'crocks/pointfree/merge'
