+++ src/Pair/fanout.js
