+++ src/Pair/fanout.spec.js
