+++ src/helpers/index.js
@@ -10,7 +10,7 @@ module.exports = {
   defaultProps: require('./defaultProps'),
   defaultTo: require('./defaultTo'),
   dissoc: require('./dissoc'),
-  fanout: require('./fanout'),
+  fanout: require('../Pair/fanout'),
   fromPairs: require('./fromPairs'),
   liftA2: require('./liftA2'),
   liftA3: require('./liftA3'),
