+++ src/helpers/index.spec.js
@@ -13,7 +13,7 @@ const curry = require('./curry')
 const defaultProps  = require('./defaultProps')
 const defaultTo  = require('./defaultTo')
 const dissoc = require('./dissoc')
-const fanout = require('./fanout')
+const fanout = require('../Pair/fanout')
 const fromPairs = require('./fromPairs')
 const liftA2 = require('./liftA2')
 const liftA3 = require('./liftA3')
