+++ src/index.spec.js
@@ -43,7 +43,7 @@ const curry = require('./helpers/curry')
 const defaultProps  = require('./helpers/defaultProps')
 const defaultTo  = require('./helpers/defaultTo')
 const dissoc = require('./helpers/dissoc')
-const fanout = require('./helpers/fanout')
+const fanout = require('./Pair/fanout')
 const fromPairs = require('./helpers/fromPairs')
 const liftA2 = require('./helpers/liftA2')
 const liftA3 = require('./helpers/liftA3')
